타입 일치 및 코드 스타일 리팩토링 적용
타입 캐스팅 일관성 강화, 상수 정의 방식 개선, 빌드 경로 및 네임스페이스 수정, 헤더 인클루드 정리, SceneManager 등 주요 코드 스타일 통일을 적용했습니다. ImGui 및 수학 함수 타입 일치, 불필요한 공백/줄바꿈 제거 등으로 코드 품질과 유지보수성을 향상시켰으며, 기능적 변화 없이 잠재적 버그 예방과 가독성 개선이 목적입니다.

+++ Client/Client/Mesh.cpp
@@ -44,26 +44,27 @@ void Mesh::upload_to_gpu_internal(ID3D12Device* device, ID3D12GraphicsCommandLis
 	// --- 기존 생성자에 있던 GPU 버퍼 생성 로직이 여기로 이전 ---
 
 	// 정점 버퍼 생성
-	_vertexBuffer = ::CreateBufferResource(device, commandList, _vertexDataBuffer.data(), _vertexDataBuffer.size(),
+	_vertexBuffer = ::CreateBufferResource(device, commandList, _vertexDataBuffer.data(), static_cast<UINT>(_vertexDataBuffer.size()),
 		D3D12_HEAP_TYPE_DEFAULT, D3D12_RESOURCE_STATE_VERTEX_AND_CONSTANT_BUFFER, &_vertexUploadBuffer);
 
 	// 인덱스 버퍼 생성 (인덱스가 있는 경우)
 	if (!_indices.empty())
 	{
-		_indexBuffer = ::CreateBufferResource(device, commandList, _indices.data(), sizeof(UINT) * _indices.size(),
+		_indexBuffer = ::CreateBufferResource(device, commandList, _indices.data(), 
+			sizeof(UINT) * static_cast<UINT>(_indices.size()),
 			D3D12_HEAP_TYPE_DEFAULT, D3D12_RESOURCE_STATE_INDEX_BUFFER, &_indexUploadBuffer);
 	}
 
 	// 버퍼 뷰 설정
 	_vertexBufferView.BufferLocation = _vertexBuffer->GetGPUVirtualAddress();
 	_vertexBufferView.StrideInBytes = _vertexStride;
-	_vertexBufferView.SizeInBytes = _vertexDataBuffer.size();
+	_vertexBufferView.SizeInBytes = static_cast<UINT>(_vertexDataBuffer.size());
 
 	if (!_indices.empty())
 	{
 		_indexBufferView.BufferLocation = _indexBuffer->GetGPUVirtualAddress();
 		_indexBufferView.Format = DXGI_FORMAT_R32_UINT;
-		_indexBufferView.SizeInBytes = sizeof(UINT) * _indices.size();
+		_indexBufferView.SizeInBytes = sizeof(UINT) * static_cast<UINT>(_indices.size());
 	}
 
 	// 이거 중요!
@@ -106,7 +107,7 @@ void Mesh::render(ID3D12GraphicsCommandList* commandList)
 	if (!_indices.empty())
 	{
 		commandList->IASetIndexBuffer(&_indexBufferView);
-		commandList->DrawIndexedInstanced(_indices.size(), 1, 0, 0, 0);
+		commandList->DrawIndexedInstanced(static_cast<UINT>(_indices.size()), 1, 0, 0, 0);
 	}
 	else
 	{
